+++ src/main/java/in/nareshit/raghu/service/impl/SpecializationServiceImpl.java
@@ -61,7 +61,7 @@ public class SpecializationServiceImpl
 	}
 	
 	@Override
-	public Map<Integer, String> getSpecializationIdAndName() {
+	public Map<Long, String> getSpecializationIdAndName() {
 		List<Object[]> list = repo.getSpecializationIdAndName();
 		return AppUtil.convertListToMap(list);
 	}
